Added: Filter volunteers to the ones that have a car for tasks requiring transportation.
Some refactoring to introduce the `Task` class and extracted a method to resolve whether a volunteer can fulfill the task.

Also, added Jest for unit testing. Figuring out whether a volunteer can fulfill a task is an important operation and I wanted to have tests that prove it works correctly.