Alternate approach to record guarding
Abstract the removal of records we don't want to process to it's own
function, filtering out the unwanted rows, then only doing our business
logic on the rows we want. This hopefully will be clean and readable,
and more easily allow future optimizations.

It also allows us to remove the guard inside the wrapping the business
logic and un-indent that code by one level, making that section a little
more readable.