Add max-len eslint rule
The `max-len` rule was not specified, and not included in
prettier/eslint defaults for some reason; tested by adding the `max-len`
rule at agreed upon 100 columns, and seeing 2 strings exceeding that
line length in dispatch.js fail.

Mark the rule as an error, but allow it to ignore long strings and URLs;
in the case of the former, we would have to change the line wrapping of
the output text if we changed the source, and that doesn't seem worth
it. URLs aren't correct if they break to a new line so those should be
excluded as well.