Match 'other' requests
Find volunteers for "other" requests, and add a message to Slack warning
that this is a guess.

The "guess" message would probably be better suited above/below the
listed volunteers section, but adding it there would require refactoring
the data scope in disptach.js, so I opted instead for a lighter touch
for now.

Also fix lint errors.