Count volunteer's assigned tasks
This implementation gets the total amount of volunteers and sets that to
a `Map`, with the user id as the key and the count of their tasks as the
value.

This Map is then passed through the function chain to end up in dispatch
where each volunteer's id is retrieved from the Map to get the amount of
tasks assigned. This scope passing as function argument could be made
more elegant.

Also, re-work the Airtable queries to grab only the data we need,
instead of doing the filtering after the fact.

On a small note, replace the `volunteerLine` constant with a direct
`return`. While the intermediate constant is not necessary, the real
reason to do this was to make ESLint happy with max line length.