Change how request task splitting works
Instead of creating new records for each individual request, then
marking the original record as "Was split", modify the original
request's "Tasks" field to use only the first task listed, then create
new records for any subsequent tasks.

This is done because the "Was split" field often gets deleted, or
filters do no use it. This shows the original record, any it appears as
a duplicate and unwanted noise.

Update a new "Original Tasks" field with all the tasks, un-split, input
in the original request. Add these tasks to each split record as well as
the original, for tracking purposes.